Add Spritesheet support and finished textured batch rendering

+++ Framework2D/Graphics/Vertex.cs
@@ -11,7 +11,7 @@ namespace Framework2D.Graphics
     {
         public Vector3 position;
         public Vector2 texCoord;
-        public int texSlot;
+        public float texSlot;
 
         public static int Size
         {
